[TASK] Refactoring VolatileInstaller to Service + CommandController
