+++ Classes/Service/EnterpriseLevelEnumeration.php
@@ -1,0 +1,17 @@
+<?php
+namespace FluidTYPO3\Site\Service;
+
+/*
+ * This file is part of the FluidTYPO3/Vhs project under GPLv2 or later.
+ *
+ * For the full copyright and license information, please read the
+ * LICENSE.md file that was distributed with this source code.
+ */
+
+class EnterpriseLevelEnumeration {
+	const BY_DEFAULT = 'default';
+	const MINIMALIST = 'minimalist';
+	const SMALL = 'small';
+	const MEDIUM = 'medium';
+	const LARGE = 'large';
+}
